feat: Make subscription scheduler invalid interval a metric not warning
This gets logged all the time but it's considered somewhat normal. It's
about to get a lot more common if we move to a "received" timestamp
instead of the timestamp from the Kafka topic which is largely monotonic.
So let's remove the noisy.
