+++ tests/subscriptions/test_combined_scheduler_executor.py
@@ -18,10 +18,10 @@ from snuba.subscriptions.combined_scheduler_executor import (
 )
 from snuba.subscriptions.data import PartitionId, SubscriptionData
 from snuba.subscriptions.store import RedisSubscriptionDataStore
+from snuba.subscriptions.types import Interval
 from snuba.subscriptions.utils import Tick
 from snuba.utils.streams.configuration_builder import build_kafka_producer_configuration
 from snuba.utils.streams.topics import Topic as SnubaTopic
-from snuba.utils.types import Interval
 from tests.backends.metrics import TestingMetricsBackend
 
 redis_client = get_redis_client(RedisClientKey.SUBSCRIPTION_STORE)
